JTable Interaction
Implemented ActionListener to create a custom class to be used with the Select button on a table. This ActionListener will output the first column of the selected row, intended as the Document ID, and displays a pop up warning when no cell or row is selected. This pop up is a self contained error handling.

Added Column storability to table.