UI Change and Feature Additions
Users are now prompted to enter a database name and collection name before being shown the Documents within the respective collection.

Frames are hidden after use to retain the ability to terminate the program at any point by clicking the exit button on the menu bar.

BrowserConnection.java was created to facilitate user input prior to connecting to a database.

BrowserDocument misused JTextArea and was changed to JLabel.

Changes to casting of ObjectID and Object from reverting to a previous commit.