Database Connection UX Change
Database connection section of the UX now features a dynamically populated drop down menu listing the available Databases and the respective collections in each.

The previous database UX class, renamed to BrowserConnection2, is still available. BrowserConnection2 requires the user to enter the exact name of the database and collection to connect to. This allows for a connection to a non-existing database and collection. This can be further expanded to allow for creation of new databases and collections.

ActionConnectDatabase now takes Strings directly rather than indirectly from JComponents passed through the method. These strings are used to create a connection to a specific collection.